Create matlab_changeOrder
This code reads and creates a new set of .csv files of the data we have for training and testing. ICU length of stay is the value that's located away from other demographics data so this will create new sets of .csv files that ordered in vital signs, lab values, demographics and outcome. This is a matlab file.